backend/requirements.txt
google-adk>=1.14.0
google-genai>=1.14.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
websockets>=13.0
